Change internal character type size to 32 bit
Summary:
Currently Unicode uses 0x10FFFF code points. It is possible to represent
all of them with up to two 16 bit values (UTF-16), but this makes it
harder to e.g. check for their width.

Some test cases were changed. Originally they were added with an
assumption that the code point will be truncated to 16 bit value and
as a result changed to another code point.

Test Plan:
* All code points <= 0xFFFF should work as before
  * Start the same tmux session in two Konsoles
  * Change background to fully transparent in one of them and put it in
    front of the other one, so that all text will overlap
  * Generate characters with (you can pipe it to `fold -s | less -r`):
```
perl -XCSDL -e 'print map{chr($_), " "} 1..0xffff'
```
  * Compare output visually.

* Code points > 0xFFFF should not be truncated to 16 bits
  * "𝐀" and "퐀" should be different characters

* Some code points > 0xFFFF should have single width
  * Vertical lines below should align:
```
|𝐀 |
|𠀋|
```

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: konsole-devel, hindenburg, #konsole

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D12236

+++ src/Character.h
@@ -57,7 +57,7 @@ const RenditionFlags RE_OVERLINE       = (1 << 10);
  * detailed too be drawn cleanly at normal font scales without anti
  * -aliasing, so those are drawn as regular characters.
  */
-inline bool isSupportedLineChar(quint16 codePoint)
+inline bool isSupportedLineChar(uint codePoint)
 {
     return (codePoint & 0xFF80) == 0x2500 // Unicode block: Mathematical Symbols - Box Drawing
            && !(0x2504 <= codePoint && codePoint <= 0x250B); // Triple and quadruple dash range
@@ -82,7 +82,7 @@ public:
      * @param _real Indicate whether this character really exists, or exists
      *              simply as place holder.
      */
-    explicit inline Character(quint16 _c = ' ',
+    explicit inline Character(uint _c = ' ',
                               CharacterColor  _f = CharacterColor(COLOR_SPACE_DEFAULT, DEFAULT_FORE_COLOR),
                               CharacterColor  _b = CharacterColor(COLOR_SPACE_DEFAULT, DEFAULT_BACK_COLOR),
                               RenditionFlags  _r = DEFAULT_RENDITION,
@@ -99,7 +99,7 @@ public:
      * look up the unicode character sequence in the ExtendedCharTable used to
      * create the sequence.
      */
-    quint16 character;
+    uint character;
 
     /** A combination of RENDITION flags which specify options for drawing the character. */
     RenditionFlags rendition;
