+++ src/ExtendedCharTable.cpp
@@ -33,14 +33,14 @@
 using namespace Konsole;
 
 ExtendedCharTable::ExtendedCharTable() :
-    _extendedCharTable(QHash<ushort, ushort *>())
+    _extendedCharTable(QHash<uint, uint *>())
 {
 }
 
 ExtendedCharTable::~ExtendedCharTable()
 {
     // free all allocated character buffers
-    QHashIterator<ushort, ushort *> iter(_extendedCharTable);
+    QHashIterator<uint, uint *> iter(_extendedCharTable);
     while (iter.hasNext()) {
         iter.next();
         delete[] iter.value();
@@ -50,11 +50,11 @@ ExtendedCharTable::~ExtendedCharTable()
 // global instance
 ExtendedCharTable ExtendedCharTable::instance;
 
-ushort ExtendedCharTable::createExtendedChar(const ushort *unicodePoints, ushort length)
+uint ExtendedCharTable::createExtendedChar(const uint *unicodePoints, ushort length)
 {
     // look for this sequence of points in the table
-    ushort hash = extendedCharHash(unicodePoints, length);
-    const ushort initialHash = hash;
+    uint hash = extendedCharHash(unicodePoints, length);
+    const uint initialHash = hash;
     bool triedCleaningSolution = false;
 
     // check existing entry for match
@@ -73,7 +73,7 @@ ushort ExtendedCharTable::createExtendedChar(const ushort *unicodePoints, ushort
                     triedCleaningSolution = true;
                     // All the hashes are full, go to all Screens and try to free any
                     // This is slow but should happen very rarely
-                    QSet<ushort> usedExtendedChars;
+                    QSet<uint> usedExtendedChars;
                     const SessionManager *sm = SessionManager::instance();
                     foreach (const Session *s, sm->sessions()) {
                         foreach (const TerminalDisplay *td, s->views()) {
@@ -81,8 +81,8 @@ ushort ExtendedCharTable::createExtendedChar(const ushort *unicodePoints, ushort
                         }
                     }
 
-                    QHash<ushort, ushort *>::iterator it = _extendedCharTable.begin();
-                    QHash<ushort, ushort *>::iterator itEnd = _extendedCharTable.end();
+                    QHash<uint, uint *>::iterator it = _extendedCharTable.begin();
+                    QHash<uint, uint *>::iterator itEnd = _extendedCharTable.end();
                     while (it != itEnd) {
                         if (usedExtendedChars.contains(it.key())) {
                             ++it;
@@ -100,7 +100,7 @@ ushort ExtendedCharTable::createExtendedChar(const ushort *unicodePoints, ushort
 
     // add the new sequence to the table and
     // return that index
-    auto buffer = new ushort[length + 1];
+    auto buffer = new uint[length + 1];
     buffer[0] = length;
     for (int i = 0; i < length; i++) {
         buffer[i + 1] = unicodePoints[i];
@@ -111,14 +111,14 @@ ushort ExtendedCharTable::createExtendedChar(const ushort *unicodePoints, ushort
     return hash;
 }
 
-ushort *ExtendedCharTable::lookupExtendedChar(ushort hash, ushort &length) const
+uint *ExtendedCharTable::lookupExtendedChar(uint hash, ushort &length) const
 {
     // look up index in table and if found, set the length
     // argument and return a pointer to the character sequence
 
-    ushort *buffer = _extendedCharTable[hash];
+    uint *buffer = _extendedCharTable[hash];
     if (buffer != nullptr) {
-        length = buffer[0];
+        length = ushort(buffer[0]);
         return buffer + 1;
     } else {
         length = 0;
@@ -126,19 +126,19 @@ ushort *ExtendedCharTable::lookupExtendedChar(ushort hash, ushort &length) const
     }
 }
 
-ushort ExtendedCharTable::extendedCharHash(const ushort *unicodePoints, ushort length) const
+uint ExtendedCharTable::extendedCharHash(const uint *unicodePoints, ushort length) const
 {
-    ushort hash = 0;
+    uint hash = 0;
     for (ushort i = 0; i < length; i++) {
         hash = 31 * hash + unicodePoints[i];
     }
     return hash;
 }
 
-bool ExtendedCharTable::extendedCharMatch(ushort hash, const ushort *unicodePoints,
+bool ExtendedCharTable::extendedCharMatch(uint hash, const uint *unicodePoints,
                                           ushort length) const
 {
-    ushort *entry = _extendedCharTable[hash];
+    uint *entry = _extendedCharTable[hash];
 
     // compare given length with stored sequence length ( given as the first ushort in the
     // stored buffer )
